Add EXTRA_BORDER around the map
Instead of having an extra border only around the right and the bottom
of the map, distribute it along all sides: now the visible area starts
at coordinates (EXTRA_BORDER, EXTRA_BORDER) of the pixmap.
This means that when dragging the map for short movements, the user is
less likely to see white borders.