Fix widget repainting
Downloaded tiles did not trigger a redraw of the correct area, so some
parts could be left unpainted.

The easiest fix is to redraw the whole window in
osm_gps_map_map_redraw(): in any case this was triggering a redraw of
all the tiles, and for each of them gtk_widget_queue_draw_area() was
being called, sometimes (for the downloaded tiles) with the wrong area.
So, substitute these redraws with a single one.
Note that the removed calls to gtk_widget_queue_draw_area() could only
be invoked in a code-path that starts with osm_gps_map_map_redraw(), so
we can remove them safely.