Makefile.am: EXTRA_DIST is here totally useless
osmgpsmap-1.0.pc.in is disted because listed in AC_CONFIG_FILES
m4 files are always disted
gtk-doc.make is disted as a dependency of docs/refeence/Makefile.in
openstreetmap_gps_map_doc_DATA does not exist

EXTRA_DIST must be used only if nothing else is usable. Most of the
time, dist_* variants of variables (like *_DATA) are available.

Signed-off-by: Quentin Glidic <sardemff7+git@sardemff7.net>