OSM forward all http resquest to https.
After asking the OSM support, I was told OSM forward all http request to
https when we use .org url.
So it should work in all cases.

Trying for all urls, I saw that maps-for-free.com does the same thing.

This has an energy cost, so it would be good for the planet to change
http to https.