pkg-config: fix requirement propagation
public requires are intended for libraries that
are being included by public headers.
`libsoup` is only indluded in `widget.c`,
and thus does not need a public require.

Adding a public require makes dependents unable to properly include this.
Specifically, `cmake` (and potentially different tools)
will only pick up on `osm-gps-map` if all the dependencies it defines are there.
While it is possible to define libsoup in all dependents, this is unnecessary.

This has caused confusion downstream in nixpkgs [1] after removing libsoup from darktable.

[1] https://github.com/NixOS/nixpkgs/issues/402543